2.57 show_double fix
param type changed to (double), from (short)

+++ 2/show_data.c
@@ -17,7 +17,7 @@ void show_long(long data) {
     show_bytes((byte_pointer) &data, sizeof(long));
 }
 
-void show_double(short data) {
+void show_double(double data) {
     show_bytes((byte_pointer) &data, sizeof(double));
 }
